# docs/openapi.yaml
openapi: 3.0.0
info:
  title: CMart Cloud-Native E-Commerce Platform API
  description: |
    Canonical REST API contract specifications for the CMart microservices suite.
    All public endpoints are versioned and follow identical payload wrapper formats and standardized error responses.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
    description: Auth Service
  - url: http://localhost:3002/api/v1
    description: Product Service
  - url: http://localhost:3003/api/v1
    description: Cart Service
  - url: http://localhost:3004/api/v1
    description: Order Service
  - url: http://localhost:3005/api/v1
    description: Payment Service
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      summary: Retrieve products catalog
      security: []
      tags:
        - Product Service
      parameters:
        - name: searchTerm
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Catalog retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsPaginatedResponse'

    post:
      summary: Create a product (Admin only)
      tags:
        - Product Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSuccessResponse'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '403':
          $ref: '#/components/responses/403Error'
  /products/{id}:
    get:
      summary: Get product by ID
      security: []
      tags:
        - Product Service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSuccessResponse'
        '404':
          $ref: '#/components/responses/404Error'

    put:
      summary: Update a product (Admin only)
      tags:
        - Product Service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSuccessResponse'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '403':
          $ref: '#/components/responses/403Error'
        '404':
          $ref: '#/components/responses/404Error'

    delete:
      summary: Delete a product (Admin only)
      tags:
        - Product Service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product soft-deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessNoDataResponse'
        '401':
          $ref: '#/components/responses/401Error'
        '403':
          $ref: '#/components/responses/403Error'
        '404':
          $ref: '#/components/responses/404Error'

  # ---------------------------------------------------------------------------
  # Cart Service Paths
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    400Error:
      description: Validation/Request details invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            success: false
            message: "Validation failed"
            code: "VALIDATION_ERROR"
            timestamp: "2026-07-17T14:06:30Z"
            requestId: "f4db7d55-8db6-4444-be1f-110cf9df28e2"
            details:
              email: "Invalid email format"
    401Error:
      description: Access token missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            success: false
            message: "Authentication token is required"
            code: "UNAUTHORIZED"
            timestamp: "2026-07-17T14:06:30Z"
            requestId: "c552a8a5-d8aa-41f8-9a3d-4cbf301e12ea"
            details: null
    403Error:
      description: Unauthorized client permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 403
            success: false
            message: "You do not have permission to access this resource"
            code: "FORBIDDEN"
            timestamp: "2026-07-17T14:06:30Z"
            requestId: "1ccf8dfc-dbd8-4f11-bef2-efdfefcf3d3c"
            details: null
    404Error:
      description: Requested item not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            success: false
            message: "Resource not found"
            code: "NOT_FOUND"
            timestamp: "2026-07-17T14:06:30Z"
            requestId: "2fdc0cd5-9dbd-424a-9eb8-bcbfecdf78de"
            details: null
...
  schemas:
...
    CreateProductRequest:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        stock:
          type: integer
          minimum: 0
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        stock:
          type: integer
          minimum: 0
...
    SuccessNoDataResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    ProductSuccessResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProductDto'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    ProductsPaginatedResponse:
      type: object
      required:
        - success
        - data
        - pagination
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductDto'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
        timestamp:
          type: string
          format: date-time
...
    ProductDto:
      type: object
      required:
        - id
        - name
        - description
        - price
        - stock
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    PaginationMetadata:
      type: object
      required:
        - page
        - limit
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer

    # Base Error Response Schema
    ErrorResponse:
      type: object
      required:
        - status
        - success
        - message
        - code
        - timestamp
        - requestId
        - details
      properties:
        status:
          type: integer
          description: HTTP status code
        success:
          type: boolean
          example: false
        message:
          type: string
          description: User-friendly error message description
        code:
          type: string
          description: Standardized error code identifier
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
          nullable: true
          description: Unique correlation request tracking ID
        details:
          type: object
          nullable: true
          description: Additional validation breakdown or specific field failure descriptions